+++ src/OrchardCore.Modules/OrchardCore.Layers/Controllers/AdminController.cs
@@ -20,6 +20,7 @@ using OrchardCore.Layers.Models;
 using OrchardCore.Layers.Services;
 using OrchardCore.Layers.ViewModels;
 using OrchardCore.Rules;
+using OrchardCore.Rules.Services;
 using OrchardCore.Settings;
 using YesSql;
