+++ src/OrchardCore.Modules/OrchardCore.Layers/Controllers/LayerRuleController.cs
@@ -12,6 +12,7 @@ using OrchardCore.DisplayManagement.Notify;
 using OrchardCore.Layers.Services;
 using OrchardCore.Layers.ViewModels;
 using OrchardCore.Rules;
+using OrchardCore.Rules.Services;
 
 namespace OrchardCore.Layers.Controllers
 {
