+++ src/OrchardCore.Modules/OrchardCore.Layers/Services/LayerFilter.cs
@@ -16,6 +16,7 @@ using OrchardCore.Layers.Models;
 using OrchardCore.Layers.ViewModels;
 using OrchardCore.Mvc.Utilities;
 using OrchardCore.Rules;
+using OrchardCore.Rules.Services;
 
 namespace OrchardCore.Layers.Services
 {
