+++ src/OrchardCore/OrchardCore.Rules.Abstractions/Services/IConditionResolver.cs
