+++ src/OrchardCore/OrchardCore.Rules.Abstractions/Services/IRuleService.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 
-namespace OrchardCore.Rules
+namespace OrchardCore.Rules.Services
 {
     public interface IRuleService
     {
